# backend/swagger.yaml
openapi: 3.0.0
info:
  title: PCMaxing API
  description: Comprehensive API Documentation for PCMaxing
  version: 1.0.0
servers:
  - url: http://localhost:5000/api
    description: Local Dev Server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
  /auth/signup:
    post:
      summary: Register a new user
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
                username: { type: string }
                password: { type: string }
      responses:
        '200': { description: "User created" }
  /auth/login:
    post:
      summary: Login user
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        '200': { description: "Logged in successfully" }
...
  /components:
    get:
      summary: List components
      tags: [Components]
      security: []
      parameters:
        - name: category
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
        - name: sort
          in: query
          schema: { type: string, enum: [price_asc, price_desc] }
      responses:
        '200': { description: "List of components" }
  /components/{id}:
    get:
      summary: Get component by ID
      tags: [Components]
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: "Component details" }
...